refactor class out, add conversion in Enso side
