+++ std-bits/table/src/main/java/org/enso/table/problems/AggregatedProblems.java
@@ -33,6 +33,10 @@ public class AggregatedProblems {
     return problems.toArray(Problem[]::new);
   }
 
+  public List<Problem> asList() {
+    return problems;
+  }
+
   public int getCount() {
     return count;
   }
