+++ std-bits/table/src/main/java/org/enso/table/problems/WithAggregatedProblems.java
@@ -1,0 +1,5 @@
+package org.enso.table.problems;
+
+/**  A value annotated with aggregated problems that occurred when it was being computed. */
+public record WithAggregatedProblems<T>(T value, AggregatedProblems problems) {
+}
